Match a bunch of overlay draw functions (#2034)
* Match a bunch of actor draw functions

* Change z_demo_gt.c draw functions to take Actor* thisx

+++ src/overlays/actors/ovl_En_Zo/z_en_zo.c
@@ -172,7 +172,7 @@ void EnZo_UpdateEffectsSplashes(EnZo* this) {
 void EnZo_DrawEffectsRipples(EnZo* this, PlayState* play) {
     EnZoEffect* effect;
     s16 i;
-    u8 materialFlag;
+    s16 materialFlag;
 
     effect = this->effects;
     OPEN_DISPS(play->state.gfxCtx, "../z_en_zo_eff.c", 217);
